Simplify dispatcher factory
This slightly reduces the complexity of the dispatcher factory, by
removing the broken inheritance. Before the prototype was manually swapped
out which is confusing and slow.

Creating a new object by composition (AnonymousObservable + Dispatcher methods)
should be faster and easier to read.

Furthermore the setPrototypeOf method is experimental and not widely supported.
Thus this commit should improve the compatibility to older browsers.