Implement proof of concept agendas
Every event on the dispatcher is now an observable. This is
called an `agenda`, because that means that every event on the
dispatcher has its own tasks, actions and side effects. Conceptually
these `agendas` / observables are then merged into one and the
resulting observable containing only actions is scanned / reduced to
produce the store state.

If an observable on the dispatcher is emitting an error, the
store's states are updated and all the actions / changes that the
`agenda` did to the stores are reverted without corrupting any
other changes. The store's history is saved in a linked list,
containing `anchors`. The anchors describe the store states and
actions. Thus it is inexpensive to recompute the store's state at
any time.

The goal of this chnge is perfect side-effect management, while
preserving the previous API.

The new method `schedule` was added, that emits observables on
the dispatcher.

If you're observing the dispatcher directly, you'll need to use
`.mergeAll()` to resolve the `agendas` / observables into actual
actions.